moved to tuya folder, created enchanted dev class
